Fix XM effects and finetuned c2spd
